+++ src/lib.rs
@@ -189,8 +189,8 @@ pub use crate::format::{
     ValueFormatBoolean, ValueFormatCurrency, ValueFormatDateTime, ValueFormatNumber,
     ValueFormatPercentage, ValueFormatRef, ValueFormatText, ValueFormatTimeDuration,
 };
-pub use crate::io::read::{read_ods, read_ods_buf};
-pub use crate::io::write::{write_ods, write_ods_buf, write_ods_buf_uncompressed};
+pub use crate::io::read::{read_ods, read_ods_buf, read_ods_from};
+pub use crate::io::write::{write_ods, write_ods_buf, write_ods_buf_uncompressed, write_ods_to};
 pub use crate::refs::{CellRange, CellRef, ColRange, RowRange};
 pub use crate::style::units::{Angle, Length};
 pub use crate::style::{CellStyle, CellStyleRef};